+++ tests/tensor/test_swap_gate.py
@@ -152,15 +152,18 @@ def test_swap_gate_charge(config_kwargs):
                         [cuuh, cndcdh, cdpdm]]:
 
             xyz0 = yastn.ncon([x, y, z], [(-0, -1), (-2, -3), (-4, -5)])
-            xyz0 = xyz0.swap_gate(axes=(0, 2), charge=z.n)
+            xyz0 = xyz0.swap_gate(axes=(0, 0, 2), charge=(y.n, z.n, z.n))
 
             x1 = x.swap_gate(axes=0, charge=z.n)
+            x1 = x1.swap_gate(axes=0, charge=y.n)
             y1 = y.swap_gate(axes=0, charge=z.n)
             xyz1 = yastn.ncon([x1, y1, z], [(-0, -1), (-2, -3), (-4, -5)])
 
             z = z.add_leg(axis=2)
-            xyz2 = yastn.ncon([x, y, z], [(-0, -1), (-2, -3), (-4, -5, -6)])
-            xyz2 = xyz2.swap_gate(axes=((0, 2), 6)).remove_leg(axis=6)
+            y = y.add_leg(axis=2)
+            xyz2 = yastn.ncon([x, y, z], [(-0, -1), (-2, -3, -6), (-4, -5, -7)])
+            xyz2 = xyz2.swap_gate(axes=((0, 2), 7)).remove_leg(axis=7)
+            xyz2 = xyz2.swap_gate(axes=(0, 6)).remove_leg(axis=6)
 
             assert (xyz0 - xyz2).norm() < tol
             assert (xyz1 - xyz2).norm() < tol
@@ -175,7 +178,7 @@ def test_swap_gate_exceptions(config_kwargs):
                        match='Odd number of elements in axes. Elements of axes should come in pairs.'):
         a.swap_gate(axes=(0, 1, 2))
     with pytest.raises(yastn.YastnError,
-                       match=r'Length of charge \(1, 1\) does not match sym.NSYM = 1.'):
+                       match=r'Length or number of charges does not match sym.NSYM or axes.'):
         a.swap_gate(axes=(0,), charge=(1, 1))